Adjust logging levels and move logging to own module
